fix(aspect): the wrap-tolerance test, not the float32 output, was the defect
359.99999263 deg and 0.0 deg are 7.4e-6 degrees apart on the circle and
360 apart on the number line. test_circular_mean_survives_the_wrap...
compared two bearings with math.isclose(..., abs_tol=1e-3) -- a linear
tolerance across the 0/360 seam -- which is exactly the class of bug
issue #201 exists to fix, reproduced inside the test meant to prove it
fixed. The prior commit's float64 cast masked the symptom (shrinking the
per-cell trig residual until the linear assertion could no longer see it)
without fixing the assertion, and cost ~50% more peak memory on gfv2's
worst batch for precision the on-disk float32 rasters never carried.

Revert run_aspect_batch to the brief's verbatim float32 radians line. Add
a _circular_deg_apart() helper to the test and switch both bearing
assertions to compare on the circle instead of linearly.

+++ src/gfv2_params/zonal_runners/aspect.py
@@ -168,15 +168,7 @@ def run_aspect_batch(config: dict, batch_id: int, logger) -> None:
     # `slope == 0` is False where slope is NaN, so nodata never counts as flat.
     flat = slope_da == FLAT_SLOPE
     sloped_aspect = aspect_da.where(~flat)
-    # Cast to float64 before the trig ops: the CONUS aspect/slope tiles are
-    # float32 (~7 significant digits), and sin/cos of a value near the 0/360
-    # wrap can leave a ~1e-7 per-cell residual -- large enough that an exactly
-    # symmetric HRU's mean_sin lands a hair on the negative side of zero rather
-    # than at literal 0.0, wrapping through atan2_deg's `% 360` to ~359.99999
-    # instead of ~0.00001 (both the same bearing, but math.isclose(..., 0.0)
-    # cannot tell that). float64 keeps the residual within the ~1e-13 ULP noise
-    # atan2_deg's double modulo is built to absorb.
-    radians = np.deg2rad(sloped_aspect.astype("float64"))
+    radians = np.deg2rad(sloped_aspect)
     # Arithmetic on a DataArray keeps coords but the .rio accessor needs the CRS
     # restated before UserTiffData reads it back off the array.
     sin_da = np.sin(radians).rio.write_crs(aspect_da.rio.crs)
